feat/todo-app#33: modify hook for register
